move special battles to own files
