+++ asm/macros/event.inc
