+++ include/battle_tower.h
@@ -20,24 +20,14 @@ extern const u8 gTowerMaleFacilityClasses[30];
 extern const u16 gTowerMaleTrainerGfxIds[30];
 extern const u8 gTowerFemaleFacilityClasses[20];
 extern const u16 gTowerFemaleTrainerGfxIds[20];
-extern const struct TrainerMon gBattleFrontierMons[];
-extern const struct BattleFrontierTrainer gBattleFrontierTrainers[];
 extern const struct TrainerMon gSlateportBattleTentMons[];
 extern const struct BattleFrontierTrainer gSlateportBattleTentTrainers[];
 
 void CallBattleTowerFunc(void);
-void SetEReaderTrainerGfxId(void);
 void PutNewBattleTowerRecord(struct EmeraldBattleTowerRecord *newRecordEm);
-void DoSpecialTrainerBattle(void);
 void CalcEmeraldBattleTowerChecksum(struct EmeraldBattleTowerRecord *record);
 void CalcRubyBattleTowerChecksum(struct RSBattleTowerRecord *record);
 u16 GetCurrentBattleTowerWinStreak(u8 lvlMode, u8 battleMode);
-u8 GetEreaderTrainerFrontSpriteId(void);
-u8 GetEreaderTrainerClassId(void);
-void GetEreaderTrainerName(u8 *dst);
-void ValidateEReaderTrainer(void);
-void ClearEReaderTrainer(struct BattleTowerEReaderTrainer *ereaderTrainer);
-void CopyEReaderTrainerGreeting(void);
 void TryHideBattleTowerReporter(void);
 bool32 RubyBattleTowerRecordToEmerald(struct RSBattleTowerRecord *src, struct EmeraldBattleTowerRecord *dst);
 bool32 EmeraldBattleTowerRecordToRuby(struct EmeraldBattleTowerRecord *src, struct RSBattleTowerRecord *dst);
