+++ include/constants/battle_special.h
@@ -1,0 +1,10 @@
+#ifndef GUARD_CONSTANTS_BATTLE_SPECIAL_H
+#define GUARD_CONSTANTS_BATTLE_SPECIAL_H
+
+// Ids for special multi battle types
+#define MULTI_BATTLE_2_VS_2 0
+#define MULTI_BATTLE_2_VS_WILD 1
+#define MULTI_BATTLE_2_VS_1 2
+#define MULTI_BATTLE_CHOOSE_MONS 0x80
+
+#endif // GUARD_CONSTANTS_BATTLE_SPECIAL_H
