+++ src/mystery_event_script.c
@@ -1,6 +1,6 @@
 #include "global.h"
 #include "berry.h"
-#include "battle_tower.h"
+#include "battle_special.h"
 #include "easy_chat.h"
 #include "event_data.h"
 #include "mail.h"
